Fix Wasm build

+++ core/silkworm/rlp/encode.hpp
@@ -39,7 +39,7 @@ namespace rlp {
 
     struct Header {
         bool list{false};
-        uint64_t payload_length{0};
+        size_t payload_length{0};
     };
 
     constexpr uint8_t kEmptyStringCode = 0x80;
